+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+- Scan jobs: Correct resolution order for options (CLI, then scanjob config, then environment variables).
+
 ## 0.4.0 (11 December 2025)
 
 - Initial release.